Fix regression in domain to session type mapping
If no domain is passed, self->session->client_info->domain is an empty
string that is incorrectly treated as if the domain were passed by the
client.

The regression was introduced in fc753a95, when an incorrect check for
non-empty domain was added. That check was removed in 67119ecc based on
coverity report, leaving no trace of the original intention.

Check domain both for '_' and '\0' as the initial character. In either
case, select the first session type section in xrdp.ini.