waitforx: remove unneded header
The sys/signal.h is essentially a redirect to signal.h, without further
logic. However when using musl libc, a warning (which is treated as an error)
is also emitted about this:

| /usr/include/sys/signal.h:1:2: error: #warning redirecting incorrect #include <sys/signal.h> to <signal.h> [-Werror=cpp]
|     1 | #warning redirecting incorrect #include <sys/signal.h> to <signal.h>

Remove sys/signal.h header to solve the problem - the signal.h header is already
included in this file.

Signed-off-by: Gyorgy Sarvari <skandigraun@gmail.com>